add: testNewFreestyleProjectCreated()
